+++ examples/.v_env/bin/django-admin.py
@@ -1,0 +1,5 @@
+#!/home/amoki/forks/django-websocket-redis/examples/.v_env/bin/python
+from django.core import management
+
+if __name__ == "__main__":
+    management.execute_from_command_line()